+++ webgl_generator/lib.rs
@@ -16,7 +16,14 @@
 //! used to generate all constants and functions of a given OpenGL version.
 //!
 //! See the `webgl` crate for an example of use.
+extern crate webidl;
+extern crate heck;
+extern crate khronos_api;
+extern crate xml;
+extern crate regex;
+extern crate html2runes;
 
+mod utils;
 mod webgl_generators;
 mod webgl_registry;
